メモ機能の責務をMemoContainerに移管
